refactor(hashtower): merge checkParameter.ts types/index.ts -> hashtower-hash-chain-proof-builder.ts
